+++ mobile/lib/app/app.dart
@@ -1,4 +1,5 @@
 import 'package:curitalk/app/router/app_router.dart';
+import 'package:curitalk/app/theme/app_theme.dart';
 import 'package:flutter/material.dart';
 
 class CuritalkApp extends StatelessWidget {
@@ -9,6 +10,7 @@ class CuritalkApp extends StatelessWidget {
     return MaterialApp.router(
       title: 'Curitalk',
       debugShowCheckedModeBanner: false,
+      theme: AppTheme.light,
       routerConfig: appRouter,
     );
   }